requirements.txt
numpy==2.0.0
numpy-stl==3.1.2
opencv-python==4.10.0.84
tqdm==4.66.5
